requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic[email]
supabase==2.18.0
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20